requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
backports.zoneinfo==0.2.1; python_version < '3.9'
blinker==1.7.0; python_version >= '3.8'
brotli==1.1.0
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
configargparse==1.7; python_version >= '3.5'
django==4.2.11; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
flask==3.0.3; python_version >= '3.8'
flask-cors==4.0.0
flask-login==0.6.3; python_version >= '3.7'
gevent==24.2.1; python_version >= '3.8'
geventhttpclient==2.0.12
greenlet==3.0.3; python_version < '3.11' and platform_python_implementation == 'CPython'
idna==3.7; python_version >= '3.5'
importlib-metadata==7.1.0; python_version < '3.10'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.3; python_version >= '3.7'
locust==2.25.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
msgpack==1.0.8; python_version >= '3.8'
psutil==5.9.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyzmq==26.0.2; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
roundrobin==0.0.4
setuptools==69.5.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.0; python_version >= '3.8'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.11.0; python_version < '3.11'
urllib3==2.2.1; python_version >= '3.8'
werkzeug==3.0.2; python_version >= '3.8'
zipp==3.18.1; python_version >= '3.8'
zope.event==5.0; python_version >= '3.7'
zope.interface==6.3; python_version >= '3.7'
